[Fix] fix eslint
